Add ability to use a temporary directory for test output
Uses the `tempfile` module to create a temporary directory for test
output, erasing it afterwards.  This patch changes the working directory
to the temporary directory to capture all output.  A new key word argument
`use_temp` controls whether this is done or not so it can still be
overridden for testing tests.